requirements.txt
streamlit>=1.28.0
google-genai==0.3.0
elevenlabs>=0.2.27
python-dotenv>=1.0.0
requests>=2.31.0
setuptools>=65.0.0
